requirements.txt
tensorflow>=2.10
pillow
numpy
pandas
scikit-learn
